fix: merge proptest_config into __tests.rs to satisfy architecture boundary checks
