+++ crates/monochange_config/src/lib.rs
@@ -4464,6 +4464,3 @@ mod mutant_killers;
 
 #[cfg(test)]
 mod __tests;
-
-#[cfg(test)]
-mod proptest_config;
